AI/requirements.txt
flask
flask-cors
pandas
requests
sentence-transformers
pymongo
python-dotenv
PyJWT